Added checks for weeks and dropdown for year
Checks to see if week exists (i.e. between 1-17 for regular season) and
if it does, whether or not a week is greater than the current week (i.e.
if it's currently week 14 and you enter 15). Also added a dropdown menu
for years 2009-2015 so I didn't have to do any checking for valid year.